requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
motor==3.6.0

# Authentication & Security
bcrypt==4.2.0
PyJWT==2.9.0
python-jose[cryptography]==3.3.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# HTTP Client
httpx==0.27.2

# File Handling
aiofiles==24.1.0

# Google Calendar Integration (Optional)
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Logging and Monitoring
python-json-logger==2.0.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0
